Send Interesting Old Move Redundantly
This was removed previously under the assumption that the redundancy was
barely ever relevant, given a good connection. Turns out that is not
true, some players with acceptable connections needed this to be able to
dodge reliably.

While rewriting the OldMove code anyway, use 24 additional bits for way
better acceleration replication (now full 16 bits per axis) and
millisecond accuracy of the delta.

The way the OldMove is selected was also changed, acceleration change is
now secondary to jumping and dodging. The last time the user jumped or
dodged will now always be preferred as the redundant OldMove.