Improve LCWeapons Compatibility
LCWeapons sometimes sets ClientUpdateTime to 5, when it wants to force
the client to send an update immediately to the server. When dead, IG+
does not use ClientUpdateTime in a way that works with this
interpretation of the variable.
Instead, updates actually get paused for 5 seconds. This leaves a lot of
time-delta for when a player respawns, which causes warping.

This change corrects the interpretation while dead.