Add Setting bEnablePingCompensatedSpawn
Previously always on, which caused respawn effects to desync from
visibility of player model. Effects have been now been synchronized as
well.

PingCompensatedSpawn means that players that respawn are not visible to
other players until the respawn notification has made a full round-trip
between server and client (ie. the server knows that the client can
move).
Without this players will stand on the spawn point visible to other
players, but unmoving for the duration of the initial round-trip.