Separate Body And Head
Previously, body and head overlapped. Shots from any direction could hit
the head and only if they missed would they be counted as body shots.

This changes that concept, body and head no longer overlap. Instead the
pawns hitbox is split into 2 cylinders, one for the body and one for the
head.
The arrangement of those cylinders is fixed such that the body
cylinder occupies everything from feet until the head cylinder starts
and has the same radius as the pawn.
The head cylinder occupies the top portion and has a total height of
2*HeadHalfHeight.

When crouching the head cylinder shrinks to nothing.

The head cylinder can also no longer be hit from everywhere. The impact
on the pawns collision cylinder must be more than 30% up from the middle
of the collision cylinder in order for a check against the head cylinder
to be performed.